fix(orama): break search/fetch-documents circular dependency
methods/search.ts held both the search() dispatcher (which imports the per-mode
search modules) and the fetchDocuments/fetchDocumentsWithDistinct helpers those
same modules import back — a circular dependency. Harmless under most bundlers,
but fatal under Metro/Expo with experimentalImportSupport (tree-shaking), where
it surfaces as 'fetchDocuments is not a function' (#961).

Extract the two helpers into methods/fetch-documents.ts and import them from
there in search-fulltext.ts / search-hybrid.ts. search.ts no longer has a
back-edge, so the cycle is eliminated for every bundler — no extra build step
or shipped bundle artifact. Pure code move, no behavioral change (1326 orama
tests pass).

Addresses #961. Supersedes #963, which worked around the cycle with an esbuild bundle.
